Consolidate result.go in output.go
